preview: fix rpath for pwsh
Powershell's dependencies use dlopen() to open various libraries.
As dlopen() will account for rpath, we need to ensure that all
libraries that need to be loaded by powershell are included.

Here we add the following paths to rpath:
- $ORIGIN
- $ORIGIN/../../usr/lib/x86_64-linux-gnu
- $ORIGIN/../../lib/x86_64-linux-gnu

Snapcraft will also append to the path to incorporate the paths
provided by the base (core18) snap.

This fixes running on Ubuntu 20.04.

Signed-off-by: Chris Patterson <chris.patterson@canonical.com>